Refactor code to support multiple platforms (#405)
* Refactor code to support multiple platforms

- Directory structure updated
- Forward method added to allow searches by platform

* Minor: Fix App Metadata
