+++ aws/aws_connector/aws_connector.go
