+++ aws/svc/ip_fuzzing/models/aws_fqdn_regex_map/aws_fqdn_regex_map.go
