+++ main.go
@@ -11,9 +11,10 @@ import (
 	"github.com/rollbar/rollbar-go"
 	log "github.com/sirupsen/logrus"
 
-	awsconnector "github.com/magneticstain/ip-2-cloudresource/aws_connector"
+	awsconnector "github.com/magneticstain/ip-2-cloudresource/aws/aws_connector"
+	awssearch "github.com/magneticstain/ip-2-cloudresource/aws/search"
+	gcpsearch "github.com/magneticstain/ip-2-cloudresource/gcp/search"
 	"github.com/magneticstain/ip-2-cloudresource/resource"
-	"github.com/magneticstain/ip-2-cloudresource/search"
 	"github.com/magneticstain/ip-2-cloudresource/utils"
 )
 
@@ -76,18 +77,45 @@ func outputResults(matchedResource resource.Resource, networkMapping bool, silen
 	}
 }
 
-func runCloudSearch(ipAddr string, cloudSvc string, ipFuzzing bool, advIPFuzzing bool, orgSearch bool, orgSearchXaccountRoleARN string, orgSearchRoleName string, orgSearchOrgUnitID string, networkMapping bool, silent bool, jsonOutput bool) {
-	// cloud connections
+func runAwsSearch(ipAddr string, cloudSvc string, ipFuzzing bool, advIPFuzzing bool, orgSearch bool, orgSearchXaccountRoleARN string, orgSearchRoleName string, orgSearchOrgUnitID string, networkMapping bool) (resource.Resource, error) {
 	log.Debug("generating AWS connection")
 	ac, err := awsconnector.New()
 	if err != nil {
 		log.Fatal(err)
 	}
 
+	searchCtlr := awssearch.Search{AWSConn: ac, IpAddr: ipAddr}
+
 	// search
-	log.Info("searching for IP ", ipAddr, " in ", cloudSvc, " service(s)")
-	searchCtlr := search.Search{AWSConn: ac, IpAddr: ipAddr}
-	matchingResource, err := searchCtlr.InitSearch(cloudSvc, ipFuzzing, advIPFuzzing, orgSearch, orgSearchXaccountRoleARN, orgSearchRoleName, orgSearchOrgUnitID, networkMapping)
+	log.Info("searching for IP ", ipAddr, " in ", cloudSvc, " AWS service(s)")
+	return searchCtlr.InitSearch(cloudSvc, ipFuzzing, advIPFuzzing, orgSearch, orgSearchXaccountRoleARN, orgSearchRoleName, orgSearchOrgUnitID, networkMapping)
+}
+
+func runGcpSearch(ipAddr, cloudSvc string) (resource.Resource, error) {
+	searchCtlr := gcpsearch.Search{IpAddr: ipAddr}
+
+	// search
+	log.Info("searching for IP ", ipAddr, " in ", cloudSvc, " GCP service(s)")
+	return searchCtlr.InitSearch(cloudSvc)
+}
+
+func runCloudSearch(platform string, ipAddr string, cloudSvc string, ipFuzzing bool, advIPFuzzing bool, orgSearch bool, orgSearchXaccountRoleARN string, orgSearchRoleName string, orgSearchOrgUnitID string, networkMapping bool, silent bool, jsonOutput bool) {
+	var matchingResource resource.Resource
+	var err error
+
+	platform = strings.ToLower(platform)
+
+	// cloud connections
+	switch platform {
+	case "aws":
+		matchingResource, err = runAwsSearch(ipAddr, cloudSvc, ipFuzzing, advIPFuzzing, orgSearch, orgSearchXaccountRoleARN, orgSearchRoleName, orgSearchOrgUnitID, networkMapping)
+	case "gcp":
+		matchingResource, err = runGcpSearch(ipAddr, cloudSvc)
+	default:
+		log.Error("invalid or unsupported platform provided ( ", platform, " )")
+		os.Exit(1)
+	}
+
 	if err != nil {
 		log.Fatal("failed to run search :: [ ERR: ", err, " ]")
 	}
@@ -101,27 +129,28 @@ func main() {
 	version := flag.Bool("version", false, "Outputs the version of IP2CR in use and exits")
 
 	// output
-	silentOutput := flag.Bool("silent", false, "If enabled, only output the results (default: false)")
-	jsonOutput := flag.Bool("json", false, "Outputs results in JSON format; implies usage of --silent flag (default: false)")
-	verboseOutput := flag.Bool("verbose", false, "Outputs all logs, from debug level to critical (default: false)")
+	silentOutput := flag.Bool("silent", false, "If enabled, only output the results")
+	jsonOutput := flag.Bool("json", false, "Outputs results in JSON format; implies usage of --silent flag")
+	verboseOutput := flag.Bool("verbose", false, "Outputs all logs, from debug level to critical")
 
 	// base
-	ipAddr := flag.String("ipaddr", "", "IP address to search for (default: 127.0.0.1)")
-	cloudSvc := flag.String("svc", "all", "Specific cloud service(s) to search. Multiple services can be listed in CSV format, e.g. elbv1,elbv2. Available services are: [all, cloudfront , ec2 , elbv1 , elbv2]  (default: all)")
+	platform := flag.String("platform", "aws", "Platform to target for IP search (e.g. aws, gcp, etc)")
+	ipAddr := flag.String("ipaddr", "", "IP address to search for (REQUIRED)")
+	cloudSvc := flag.String("svc", "all", "Specific cloud service(s) to search. Multiple services can be listed in CSV format, e.g. elbv1,elbv2. Available services are: [all, cloudfront , ec2 , elbv1 , elbv2]")
 
 	// FEATURE FLAGS
 	// IP fuzzing
-	ipFuzzing := flag.Bool("ip-fuzzing", true, "Toggle the IP fuzzing feature to evaluate the IP and help optimize search (not recommended for small accounts due to overhead outweighing value)  (default: true)")
-	advIPFuzzing := flag.Bool("adv-ip-fuzzing", true, "Toggle the advanced IP fuzzing feature to perform a more intensive heuristics evaluation to fuzz the service (not recommended for IPv6 addresses) (default: true)")
+	ipFuzzing := flag.Bool("ip-fuzzing", true, "Toggle the IP fuzzing feature to evaluate the IP and help optimize search (not recommended for small accounts due to overhead outweighing value)")
+	advIPFuzzing := flag.Bool("adv-ip-fuzzing", true, "Toggle the advanced IP fuzzing feature to perform a more intensive heuristics evaluation to fuzz the service (not recommended for IPv6 addresses)")
 
 	// org search
-	orgSearch := flag.Bool("org-search", false, "Search through all child accounts of the organization for resources, as well as target account (target account should be parent account) (default: false)")
-	orgSearchXaccountRoleARN := flag.String("org-search-xaccount-role-arn", "", "The ARN of the role to assume for gathering AWS Organizations information for search, e.g. the role to assume with R/O access to your AWS Organizations account (default: '')")
-	orgSearchRoleName := flag.String("org-search-role-name", "ip2cr", "The name of the role in each child account of an AWS Organization to assume when performing a search (default: ip2cr)")
-	orgSearchOrgUnitID := flag.String("org-search-ou-id", "", "The ID of the AWS Organizations Organizational Unit to target when performing a search (default: '')")
+	orgSearch := flag.Bool("org-search", false, "Search through all child accounts of the organization for resources, as well as target account (target account should be parent account)")
+	orgSearchXaccountRoleARN := flag.String("org-search-xaccount-role-arn", "", "The ARN of the role to assume for gathering AWS Organizations information for search, e.g. the role to assume with R/O access to your AWS Organizations account")
+	orgSearchRoleName := flag.String("org-search-role-name", "ip2cr", "The name of the role in each child account of an AWS Organization to assume when performing a search")
+	orgSearchOrgUnitID := flag.String("org-search-ou-id", "", "The ID of the AWS Organizations Organizational Unit to target when performing a search")
 
 	// network mapping
-	networkMapping := flag.Bool("network-mapping", false, "If enabled, generate a network map associated with the identified resource if it's found (default: false)")
+	networkMapping := flag.Bool("network-mapping", false, "If enabled, generate a network map associated with the identified resource if it's found")
 
 	flag.Parse()
 
@@ -155,7 +184,7 @@ func main() {
 
 	utils.InitRollbar(APP_ENV, APP_VER)
 
-	rollbar.WrapAndWait(runCloudSearch, *ipAddr, *cloudSvc, *ipFuzzing, *advIPFuzzing, *orgSearch, *orgSearchXaccountRoleARN, *orgSearchRoleName, *orgSearchOrgUnitID, *networkMapping, *silentOutput, *jsonOutput)
+	rollbar.WrapAndWait(runCloudSearch, *platform, *ipAddr, *cloudSvc, *ipFuzzing, *advIPFuzzing, *orgSearch, *orgSearchXaccountRoleARN, *orgSearchRoleName, *orgSearchOrgUnitID, *networkMapping, *silentOutput, *jsonOutput)
 
 	rollbar.Close()
 }
